Added Scene class to simplify drawing and updating of menus
