--- src/Game.java
@@ -1,77 +1,0 @@
-package src;
-
-import java.awt.Graphics2D;
-import java.awt.Point;
-
-/**
- * Game class that does most of the mechanics.
- * @author Harrison Fenton-Fearn
- */
-
-public class Game {
-
-
-    public Game()
-    {
-        Framework.gameState = Framework.GameState.GAME_CONTENT_LOADING;
-        
-        Thread threadForInitGame = new Thread() {
-            @Override
-            public void run(){
-                Initialize();
-                LoadContent();
-                Framework.gameState = Framework.GameState.PLAYING;
-            }
-        };
-        threadForInitGame.start();
-    }
-    
-    
-   /**
-     * Set variables and objects for the game.
-     */
-    private void Initialize()
-    {
-
-    }
-    
-    /**
-     * Load game files - images, sounds, ...
-     */
-    private void LoadContent()
-    {
-    
-    }    
-    
-    
-    /**
-     * Restart game - reset some variables.
-     */
-    public void RestartGame()
-    {
-        
-    }
-    
-    
-    /**
-     * Update game logic.
-     * 
-     * @param gameTime gameTime of the game.
-     * @param mousePosition current mouse position.
-     */
-    public void UpdateGame(long gameTime, Point mousePosition)
-    {
-        
-    }
-    
-    /**
-     * Draw the game to the screen.
-     * 
-     * @param g2d Graphics2D
-     * @param mousePosition current mouse position.
-     */
-    public void Draw(Graphics2D g2d, Point mousePosition)
-    {
-        
-    }
-}
